Made rooms vertical. added sperate classes
I made the the rooms vertical. I want to have a reference. I want it to be like ( white bar with circles inside with a slightly different color) when people join their profile pic takes the circles space) I might download clubhouse to see how it looks like. I didn't do that much because I was getting hang of the CSS and html and I didn't wanna mess up anything.